Fix jumping scrollbar issue
